# docs/cms-api/swagger.yaml
basePath: /
definitions:
...
  internal_cms_handler.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_cms_handler.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
    type: object
...
  internal_cms_handler.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  internal_cms_handler.RefreshResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Content Management System API for MediaCMS
  title: MediaCMS API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password, returns access and refresh
        tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_cms_handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_cms_handler.LoginResponse'
        "401":
          description: Invalid credentials
          schema:
            type: string
      summary: Login
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate the refresh token to logout the user
      parameters:
      - description: Refresh token to invalidate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_cms_handler.RefreshRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "401":
          description: Invalid refresh token
          schema:
            type: string
        "500":
          description: Failed to logout
          schema:
            type: string
      summary: Logout
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token for a new access token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_cms_handler.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_cms_handler.RefreshResponse'
        "401":
          description: Invalid refresh token
          schema:
            type: string
      summary: Refresh token
      tags:
      - auth
...
  /api/v1/users:
    post:
      consumes:
      - application/json
      description: Create a new user account with editor role (admin only)
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_cms_handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_cms_handler.LoginResponse'
        "400":
          description: Invalid request
          schema:
            type: string
        "409":
          description: User with this email already exists
          schema:
            type: string
        "500":
          description: Failed to create user
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"